refactor history & se elimino tabla example
