--- src/app/components/private/students/components/table-example/students-dashboard.component.html
@@ -1,4 +1,0 @@
-<!-- <p>students-dashboard works!</p>
-<app-tabla-data  [titulo]="title"
-[columns]="columnas"
-[data]="mouseList" (action)="onAction($event)"></app-tabla-data> -->
