+++ src/app/components/private/students/index.ts
@@ -1,6 +1,5 @@
 export * from './components/layout/students-layout.component';
 export * from './components/table-patients/students-patients.component';
-export * from './components/table-example/students-dashboard.component';
 export * from './components/odontogram/students-odontogram.component';
 export * from './components/toolbar-odontogram/students-toolbar.component'
 export * from './components/tooth/students-tooth.component';
