+++ src/app/components/private/students/pages/history-clinics/general/students-general-history.component.scss
