+++ src/app/components/private/students/pages/history-clinics/general/students-general-history.component.ts
@@ -2,18 +2,18 @@ import { Component, OnInit, ViewChild } from '@angular/core';
 import { MatTabGroup, MatTabsModule } from '@angular/material/tabs';
 import { ActivatedRoute } from '@angular/router';
 
-import { HistoryPersonalDataComponent } from '../../components/form-personal-data/history-personal-data.component';
-import { HistoryVitalSignsComponent } from '../../components/form-history-vital-signs/history-vital-signs.component';
+import { HistoryPersonalDataComponent } from '../../../components/form-personal-data/history-personal-data.component';
+import { HistoryVitalSignsComponent } from '../../../components/form-history-vital-signs/history-vital-signs.component';
 //import { HistoryFacialExamComponent } from './history-facial-exam/history-facial-exam.component';
 import { MatDialogModule } from '@angular/material/dialog';
-import { StudentsOdontogramComponent } from '../../components/odontogram/students-odontogram.component';
-import { HistoryFacialExamComponent } from '../../components/form-history-facial-exam/history-facial-exam.component';
-import { HistoryFamilyHistoryComponent } from '../../components/form-history-family-history/history-family-history.component';
-import { HistoryInitialBagComponent } from '../../components/form-history-initial-bag/history-initial-bag.component';
-import { HistoryMultidiciplinaryEvaluationComponent } from '../../components/form-history-multidiciplinary-evaluation/history-multidiciplinary-evaluation.component';
-import { NoPathologicalPersonalHistoryComponent } from '../../components/form-no-pathological-personal-history/no-pathological-personal-history.component';
-import { PathologicalPersonalHistoryComponent } from '../../components/form-pathological-personal-history/pathological-personal-history.component';
-import { StudentsPatientDetailComponent } from '../../components/form-patient-detail/students-patient-detail.component';
+import { StudentsOdontogramComponent } from '../../../components/odontogram/students-odontogram.component';
+import { HistoryFacialExamComponent } from '../../../components/form-history-facial-exam/history-facial-exam.component';
+import { HistoryFamilyHistoryComponent } from '../../../components/form-history-family-history/history-family-history.component';
+import { HistoryInitialBagComponent } from '../../../components/form-history-initial-bag/history-initial-bag.component';
+import { HistoryMultidiciplinaryEvaluationComponent } from '../../../components/form-history-multidiciplinary-evaluation/history-multidiciplinary-evaluation.component';
+import { NoPathologicalPersonalHistoryComponent } from '../../../components/form-no-pathological-personal-history/no-pathological-personal-history.component';
+import { PathologicalPersonalHistoryComponent } from '../../../components/form-pathological-personal-history/pathological-personal-history.component';
+import { StudentsPatientDetailComponent } from '../../../components/form-patient-detail/students-patient-detail.component';
 
 @Component({
   selector: 'app-students-general-history',
