+++ src/app/components/private/students/pages/history-clinics/oral-surgery/students-oral-surgery-history.component.ts
@@ -1,10 +1,10 @@
 import { Component, ViewChild } from '@angular/core';
 import { MatTabGroup, MatTabsModule } from '@angular/material/tabs';
-import { CurrentConditionComponent } from '../../components/form-current-condition/current-condition.component';
-import { FamilyHistoryComponent } from '../../components/form-family-history/family-history.component';
-import { IdentificationFileComponent } from '../../components/form-identification-file-oral-surgery/identification-file.component';
-import { InterrogationDivicesSystemsComponent } from '../../components/form-interrogation-divices-systems/interrogation-divices-systems.component';
-import { VitalSingsComponent } from '../../components/form-vital-sings-oral-surgery/vital-sings.component';
+import { CurrentConditionComponent } from '../../../components/form-current-condition/current-condition.component';
+import { FamilyHistoryComponent } from '../../../components/form-family-history/family-history.component';
+import { IdentificationFileComponent } from '../../../components/form-identification-file-oral-surgery/identification-file.component';
+import { InterrogationDivicesSystemsComponent } from '../../../components/form-interrogation-divices-systems/interrogation-divices-systems.component';
+import { VitalSingsComponent } from '../../../components/form-vital-sings-oral-surgery/vital-sings.component';
 
 @Component({
   selector: 'app-students-oral-surgery-history',
